+++ apps/remnant2toolkit/types/next-auth.d.ts
@@ -1,4 +1,4 @@
-import { DefaultSession, Profile as DefaultProfile } from 'next-auth'
+import { type DefaultSession, type Profile as DefaultProfile } from 'next-auth';
 
 declare module 'next-auth' {
   /**
@@ -6,13 +6,13 @@ declare module 'next-auth' {
    */
   interface Session {
     user?: {
-      id: string
-      displayName: string
-      role: string
-    } & DefaultSession['user']
+      id: string;
+      displayName: string;
+      role: string;
+    } & DefaultSession['user'];
   }
 
   interface Profile {
-    image_url: string & DefaultProfile
+    image_url: string & DefaultProfile;
   }
 }