+++ packages/ui/src/common/tooltip/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import 'tippy.js/dist/tippy.css';
 
 import Tippy, { type TippyProps } from '@tippyjs/react';