+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2025-01-02
+
+### Changed
+
+- Improved the UI for the quality build feed on the home page, particularly while items are loading.
+
 ## 2024-12-28
 
 ### Fixed